# Middleware refactoring