--- a/actor/props.go
+++ b/actor/props.go
@@ -2,17 +2,18 @@ package actor
 
 import "github.com/AsynkronIT/protoactor-go/mailbox"
 
+type InboundMiddleware func(next ActorFunc) ActorFunc
+type OutboundMiddleware func(next SenderFunc) SenderFunc
+
 // Props represents configuration to define how an actor should be created
 type Props struct {
-	actorProducer           Producer
-	mailboxProducer         mailbox.Producer
-	supervisionStrategy     SupervisorStrategy
-	middleware              []func(next ActorFunc) ActorFunc
-	outboundMiddleware      []func(next SenderFunc) SenderFunc
-	middlewareChain         ActorFunc
-	outboundMiddlewareChain SenderFunc
-	dispatcher              mailbox.Dispatcher
-	spawner                 SpawnFunc
+	actorProducer       Producer
+	mailboxProducer     mailbox.Producer
+	supervisionStrategy SupervisorStrategy
+	inboundMiddleware   []InboundMiddleware
+	outboundMiddleware  []OutboundMiddleware
+	dispatcher          mailbox.Dispatcher
+	spawner             SpawnFunc
 }
 
 func (props *Props) getDispatcher() mailbox.Dispatcher {
@@ -43,16 +44,14 @@ func (props *Props) spawn(id string, parent *PID) (*PID, error) {
 	return DefaultSpawner(id, props, parent)
 }
 
-//WithMiddleware assigns one or more middlewares to the props
-func (props *Props) WithMiddleware(middleware ...func(ActorFunc) ActorFunc) *Props {
-	props.middleware = append(props.middleware, middleware...)
-	props.middlewareChain = makeMiddlewareChain(props.middleware, localContextReceiver)
+// Assign one or more middlewares to the props
+func (props *Props) WithMiddleware(middleware ...InboundMiddleware) *Props {
+	props.inboundMiddleware = append(props.inboundMiddleware, middleware...)
 	return props
 }
 
-func (props *Props) WithOutboundMiddleware(middleware ...func(SenderFunc) SenderFunc) *Props {
+func (props *Props) WithOutboundMiddleware(middleware ...OutboundMiddleware) *Props {
 	props.outboundMiddleware = append(props.outboundMiddleware, middleware...)
-	props.outboundMiddlewareChain = makeOutboundMiddlewareChain(props.outboundMiddleware, localContextSender)
 	return props
 }