@rogeralsing 2017-04-04
Thanks!
Looking good from a conceptual perspectie. 
But I need to have a look at how this performs in terms of spawn weight.
Doing the middleware chain construction this way will add both CPU and memory pressure.