@meAmidos 2017-04-20
I've added a simple benchmark for spawning an actor. The results vary from run to run and mostly fit into the range 4500-7000 ns/op with no consistent difference between versions before and after refactoring. Here are some concrete results:

## Before
```
BenchmarkLocalContext_SpawnWithMiddleware1-4  200000 6656 ns/op
BenchmarkLocalContext_SpawnWithMiddleware2-4  200000 7306 ns/op
BenchmarkLocalContext_SpawnWithMiddleware5-4  300000 5405 ns/op
```

## After
```
BenchmarkLocalContext_SpawnWithMiddleware0-4  200000 5845 ns/op
BenchmarkLocalContext_SpawnWithMiddleware1-4  200000 6545 ns/op
BenchmarkLocalContext_SpawnWithMiddleware2-4  200000 6808 ns/op
BenchmarkLocalContext_SpawnWithMiddleware5-4  300000 5942 ns/op
```

So, basically, there is no any noticeable performance regression. I have not measured CPU and memory usage, though.

I can see an alternative way of achieving the main goal of this PR without creating a middleware chain on every spawn. However, it would require changing the signature of inbound middleware functions which is a breaking change. I am not sure if you are open to such changes.

By the way, if you _are_ open to considering breaking changes, I would rather use this possibility to ~break something else~ to revisit the whole middleware/plugin approach. But this is a big story for another PR :)