@rogeralsing 2017-04-20
We are absolutely open for discussions on how the middleware should work.
There are a few things we would like to accomplish having the middleware.

1) Testability, it makes life easier when you can test what goes in, or out of each actor.
2) Tracing, e.g. allowing us to plug in Zipkin to the actors and trace calls between them.
*) All this while still being extremely lightweight

The perf numbers looks good here.