@meAmidos 2017-04-21
It's great to know the project is open to changes. Testability is exactly the reason why I started to dig into the middleware implementation.

Is there anything I could do to facilitate the merging of this PR?