@rogeralsing 2017-04-21
Merged, thank you for the contribution!
Please join in at the gitter chat https://gitter.im/AsynkronIT/protoactor
As we try to discuss design topics there and it would be great if we can align this into .NET too.